[refactor] share driver jars loading logic
